// src/generated/models/fICADocumentType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * iPhande API
 * Township enterprise ERP, hyper-local marketplace, and merchant credibility infrastructure.
 * OpenAPI spec version: 0.1.0
 */

export type FICADocumentType = typeof FICADocumentType[keyof typeof FICADocumentType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FICADocumentType = {
  SA_ID_DOCUMENT: 'SA_ID_DOCUMENT',
  PROOF_OF_RESIDENCE: 'PROOF_OF_RESIDENCE',
  BANK_STATEMENTS_3MO: 'BANK_STATEMENTS_3MO',
  PAYSLIP_LATEST: 'PAYSLIP_LATEST',
  DRIVERS_LICENSE: 'DRIVERS_LICENSE',
} as const;